% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_to_musicassessr_vignette_folder}
\alias{copy_to_musicassessr_vignette_folder}
\title{Copy an asset to the musicassessr vignette folder}
\usage{
copy_to_musicassessr_vignette_folder(asset)
}
\arguments{
\item{asset}{}
}
\description{
Copy an asset to the musicassessr vignette folder
}
