% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melodic_recall_paper_2023_early_vs_late_thirds_long}
\alias{melodic_recall_paper_2023_early_vs_late_thirds_long}
\title{Data for looking at different sections of a melodic recall's similarity from Silas & Müllensiefen (2023)}
\format{
Dataframe.
}
\usage{
melodic_recall_paper_2023_early_vs_late_thirds_long
}
\description{
Data for looking at different sections of a melodic recall's similarity from Silas & Müllensiefen (2023)
}
\keyword{datasets}
