% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melodic_recall_paper_2023_fit1}
\alias{melodic_recall_paper_2023_fit1}
\title{EFA fit from model in Silas & Müllensiefen (2023)}
\format{
Dataframe.
}
\usage{
melodic_recall_paper_2023_fit1
}
\description{
EFA fit from model in Silas & Müllensiefen (2023)
}
\keyword{datasets}
