% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melodic_recall_paper_2023_lm.A1.2}
\alias{melodic_recall_paper_2023_lm.A1.2}
\title{Model A1.2 from Silas & Müllensiefen (2023)}
\format{
Dataframe.
}
\usage{
melodic_recall_paper_2023_lm.A1.2
}
\description{
Model A1.2 from Silas & Müllensiefen (2023)
}
\keyword{datasets}
