% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melodic_recall_paper_2023_main}
\alias{melodic_recall_paper_2023_main}
\title{Main data from Silas & Müllensiefen (2023)}
\format{
Dataframe.
}
\usage{
melodic_recall_paper_2023_main
}
\description{
Main data from Silas & Müllensiefen (2023)
}
\keyword{datasets}
