% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melodic_recall_paper_2023_melodies_with_features}
\alias{melodic_recall_paper_2023_melodies_with_features}
\title{Melodic feature data from Silas & Müllensiefen (2023)}
\format{
Dataframe.
}
\usage{
melodic_recall_paper_2023_melodies_with_features
}
\description{
Melodic feature data from Silas & Müllensiefen (2023)
}
\keyword{datasets}
